+++ src/utils/logger.ts
@@ -1,766 +1,43 @@
 /**
  * logger.ts
- * Structured Logging Utility with Error Codes
+ * Structured Logging Utility with Error Codes — barrel re-export.
  * Stores logs in chrome.storage.local with 3-day retention policy.
  *
- * @module logger
- * @requires ./piiSanitizer.js — sanitizeLogDetails uses sanitizeRegex to mask
- *   API keys and other sensitive values before log output.
- *   See also: import type { SanitizeOptions } below for graphify edge.
- */
-import { sanitizeRegex } from './piiSanitizer.js';
-import type { SanitizeOptions } from './piiSanitizer.js';
-
-// エラーコード定義（SRE/Logging改善 #8）
-export const ErrorCode = {
-    // ストレージ関連
-    STORAGE_READ_FAILURE: 'STRG_RD_001',
-    STORAGE_WRITE_FAILURE: 'STRG_WR_001',
-    STORAGE_KEY_NOT_FOUND: 'STRG_NF_001',
-    STORAGE_MIGRATION_FAILURE: 'STRG_MIG_001',
-    STORAGE_QUOTA_EXCEEDED: 'STRG_QUOTA_001',
-    MIGRATION_ROLLBACK_FAILED: 'STRG_ROLLBACK_001',
-
-    // 暗号化関連
-    CRYPTO_DECRYPTION_FAILURE: 'CRPT_DEC_001',
-    CRYPTO_ENCRYPTION_FAILURE: 'CRPT_ENC_001',
-    CRYPTO_KEY_DERIVE_FAILURE: 'CRPT_KEY_001',
-    CRYPTO_HASH_FAILURE: 'CRPT_HSH_001',
-    CRYPTO_HMAC_FAILURE: 'CRPT_HMAC_001',
-
-    // API通信関連
-    API_REQUEST_FAILURE: 'API_REQ_001',
-    API_TIMEOUT: 'API_TIM_001',
-    API_RATE_LIMIT: 'API_RL_001',
-    API_AUTH_FAILURE: 'API_AUTH_001',
-
-    // Obsidian通信関連
-    OBSIDIAN_CONNECT_FAILURE: 'OBS_CONN_001',
-    OBSIDIAN_SEND_FAILURE: 'OBS_SEND_001',
-    OBSIDIAN_RESPONSE_PARSE_FAILURE: 'OBS_PARSE_001',
-
-    // コンテンツ抽出関連
-    CONTENT_EXTRACTION_FAILURE: 'CONT_EXT_001',
-    CONTENT_TRUNCATION: 'CONT_TRUNC_001',
-
-    // PII/プライバシー関連
-    PII_DETECTION_FAILURE: 'PII_DET_001',
-    PII_REDACTION_FAILURE: 'PII_RED_001',
-    PRIVACY_MODE_VIOLATION: 'PRIV_VIOL_001',
-
-    // 入力検証関連
-    INVALID_INPUT: 'VAL_INP_001',
-    MISSING_REQUIRED_FIELD: 'VAL_REQ_001',
-
-    // 設定管理関連
-    SETTINGS_IMPORT_FAILURE: 'SET_IMP_001',
-    SETTINGS_EXPORT_FAILURE: 'SET_EXP_001',
-    SETTINGS_SIGNATURE_FAILURE: 'SET_SIG_001',
-
-    // APIキー管理関連
-    API_KEY_EXCLUDED: 'SET_AK_EXCL_001',
-    API_KEY_MERGE_CONFLICT: 'SET_AK_MRG_001',
-
-    // Trust Database関連（Phase 1）
-    TRUST_DB_INIT_FAILED: 'TRUST_INIT_001',
-    TRUST_DB_NOT_INITIALIZED: 'TRUST_NOT_INIT_001',
-    TRUST_DB_MIGRATION_FAILED: 'TRUST_MIG_001',
-    TRANCO_FETCH_FAILED: 'TRANCO_FETCH_001',
-    TRANCO_PARSE_FAILED: 'TRANCO_PARSE_001',
-    BLOOM_FILTER_ERROR: 'BLM_FLT_001',
-
-    // CSP/AIプロバイダー関連
-    UNKNOWN_AI_PROVIDER: 'CSP_AI_001',
-
-    // 汎用エラー
-    UNKNOWN_ERROR: 'UNKN_001',
-    INTERNAL_ERROR: 'INT_001',
-
-    // UI/Badge関連
-    BADGE_UPDATE_FAILED: 'UI_BADGE_001',
-
-    // Permission Manager関連（P0）
-    PERMISSION_REQUIRED: 'PERM_REQ_001'
-} as const;
-
-export type ErrorCodeValues = typeof ErrorCode[keyof typeof ErrorCode];
-
-/**
- * Template literal type documenting the structured error code pattern.
- * Most error codes follow the format: PREFIX_SUFFIX_NUMBER (e.g., STRG_RD_001).
- * Note: TypeScript template literal matching has limitations with multiple
- * underscore segments, so this serves as documentation and future constraint
- * for new error codes rather than a strict compile-time check on all existing values.
- */
-export type ErrorCodePattern = `${string}_${string}_${number}`;
-
-const LOG_STORAGE_KEY = 'sanitization_logs';
-const RETENTION_DAYS = 3;
-const MAX_LOGS = 500; // Prevent unlimited growth
-
-// 【セキュリティ強化】log sanitizationへの深度制限と循環参照保護
-const MAX_RECURSION_DEPTH = 100; // redaction.tsと整合
-const SANITIZE_RESULT = {
-  TOO_DEEP: '[SANITIZED: too deep]',
-  CIRCULAR_REF: '[SANITIZED: circular reference]'
-} as const;
-
-// 【パフォーマンス改善】バッチ書き込み用設定
-const BATCH_FLUSH_SIZE = 10; // バッファがこのサイズを超えるとフラッシュ
-const BATCH_FLUSH_ALARM_MINUTES = 1; // chrome.alarms ベースの遅延フラッシュ（Chrome の最小間隔）
-const LOGGER_ALARM_NAME = 'yasumaro-logger-flush';
-const MAX_PENDING_LOGS = 100; // バッファ上限（メモリリーク防止）
-let pendingLogs: LogEntry[] = []; // 保留中のログバッファ
-let isFlushing = false; // フラッシュ中フラグ（多重フラッシュ防止）
-
-/**
- * 【機能概要】: 環境判定関数
- * 【実装方針】: process.env.NODE_ENVでdevelopmentかどうかを判定
- * 【テスト対応】: logger-production.test.ts
- * 🟡 信頼性レベル: 黄信号（環境変数による判定は一般的なパターンによる）
- * @returns {boolean} development環境の場合はtrue
- */
-export const isDevelopment = (): boolean => {
-  // Check process.env.NODE_ENV first
-  // This takes priority because tests explicitly set this variable
-  if (typeof process !== 'undefined' && process.env) {
-    const nodeEnv = process.env.NODE_ENV;
-    // Handle null, undefined, or missing NODE_ENV as production (safe default)
-    if (nodeEnv === 'development') {
-      return true;
-    }
-    if (nodeEnv === 'production' || nodeEnv === 'test' || nodeEnv === undefined || nodeEnv === null) {
-      return false;
-    }
-  }
-  // Fall back to Vite's import.meta.env.DEV for non-test environments
-  // This only runs when process.env.NODE_ENV is not set at all
-  if (typeof import.meta !== 'undefined' && import.meta.env && import.meta.env.DEV === true) {
-    return true;
-  }
-  return false;
-};
-
-export const LogType = {
-    INFO: 'INFO',
-    WARN: 'WARN',
-    ERROR: 'ERROR',
-    SANITIZE: 'SANITIZE',
-    DEBUG: 'DEBUG'
-} as const;
-
-export type LogTypeValues = typeof LogType[keyof typeof LogType];
-
-export interface LogEntry {
-    id: string;
-    timestamp: number;
-    type: LogTypeValues;
-    message: string;
-    errorCode?: ErrorCodeValues;
-    details?: Record<string, unknown>;
-    source?: string; // ログ出力元モジュール
-    userId?: string; // ユーザー識別子（匿名化済み）
-}
-
-/**
- * 【パフォーマンス改善】保留中のログをstorageにフラッシュする
- * @param {boolean} immediate - trueの場合は即時フラッシュ（テスト用）
- */
-export async function flushLogs(_immediate: boolean = false): Promise<void> {
-    if (isFlushing || pendingLogs.length === 0) {
-        return;
-    }
-
-    // 【オフスクリーン文書対応】chrome.storage が利用できない環境（offscreen.html）では
-    // storageへの書き込みをスキップする。オフスクリーン文書からは chrome.storage に
-    // アクセスできないため、Service Worker にメッセージで転送するか console に出力する。
-    if (typeof chrome === 'undefined' || !chrome.storage) {
-        // オフスクリーン文書など chrome.storage 未対応環境: console に出力してバッファをクリア
-        for (const log of pendingLogs) {
-            console.log(`[Logger:${log.type}] ${log.message}`, log.details || '');
-        }
-        pendingLogs = [];
-        clearScheduledFlush();
-        return;
-    }
-
-    isFlushing = true;
-
-    try {
-        // バッファの内容をコピーしてクリア
-        const logsToFlush = [...pendingLogs];
-        pendingLogs = [];
-
-        // アラームスケジュールをクリア
-        clearScheduledFlush();
-
-        // 既存ログを取得
-        const storage = await chrome.storage.local.get(LOG_STORAGE_KEY);
-        let logs: LogEntry[] = storage[LOG_STORAGE_KEY] as LogEntry[] || [];
-
-        // 新しいログを追加
-        logs.push(...logsToFlush);
-
-        // 古いログを削除
-        logs = pruneLogs(logs);
-
-        // サイズ制限
-        if (logs.length > MAX_LOGS) {
-            logs = logs.slice(logs.length - MAX_LOGS);
-        }
-
-        // storageに保存
-        await chrome.storage.local.set({ [LOG_STORAGE_KEY]: logs });
-    } catch (e) {
-        console.error('Logger: Failed to flush logs', e);
-    } finally {
-        isFlushing = false;
-    }
-}
-
-/**
- * 【パフォーマンス改善】保留中のログをスケジュールしてフラッシュする
- * chrome.alarms を使用して Service Worker サスペンド後も再開可能にする。
- */
-function scheduleFlush(): void {
-    if (typeof chrome === 'undefined' || !chrome.alarms) {
-        return;
-    }
-
-    chrome.alarms.create(LOGGER_ALARM_NAME, { delayInMinutes: BATCH_FLUSH_ALARM_MINUTES });
-}
-
-/**
- * スケジュール済みのフラッシュアラームをクリアする
- */
-function clearScheduledFlush(): void {
-    if (typeof chrome === 'undefined' || !chrome.alarms) {
-        return;
-    }
-
-    chrome.alarms.clear(LOGGER_ALARM_NAME);
-}
-
-/**
- * chrome.alarms イベントで保留中ログをフラッシュする
- */
-if (typeof chrome !== 'undefined' && chrome.alarms && chrome.alarms.onAlarm) {
-    chrome.alarms.onAlarm.addListener((alarm) => {
-        if (alarm.name === LOGGER_ALARM_NAME) {
-            void flushLogs();
-        }
-    });
-}
-
-/**
- * 【Service Worker対策】サスペンド前にログを即時フラッシュ
- * ChromeがService Workerを停止する前に保留中のログを保存
- */
-if (typeof chrome !== 'undefined' && chrome.runtime && chrome.runtime.onSuspend) {
-    chrome.runtime.onSuspend.addListener(async () => {
-        console.log('[Logger] Service Worker suspending - flushing pending logs');
-        const pendingCountBeforeFlush = pendingLogs.length;
-        // Await flush with a timeout so pending logs are not lost when the SW dies.
-        const flushCompleted = await Promise.race([
-            flushLogs(true).then(() => true),
-            new Promise<boolean>((resolve) => setTimeout(() => resolve(false), 3000)),
-        ]);
-        // Best-effort visibility: chrome.storage may itself be unreliable during
-        // suspend, so we can't persist this — but it surfaces in the SW's console
-        // output (captured by chrome://extensions inspect) for debugging.
-        if (!flushCompleted) {
-            console.error(
-                `[Logger] Flush timed out during suspend — up to ${pendingCountBeforeFlush} log entries may not have been persisted`
-            );
-        }
-    });
-}
-
-/**
- * 【パフォーマンス改善】保留中のログの数を取得（テスト用）
- */
-export function getPendingLogCount(): number {
-    return pendingLogs.length;
-}
-
-/**
- * 【パフォーマンス改善】保留中のログをクリア（テスト用）
- */
-export function clearPendingLogs(): void {
-    pendingLogs = [];
-}
-
-/**
- * ログの詳細情報をサニタイズする（PII検出とマスキング）
- * 【深度制限と循環参照保護】
- * @param {Record<string, unknown>} details - サニタイズ対象の詳細情報
- * @param {WeakSet<object>} [visitedObjects] - 循環参照検出用WeakSet
- * @param {number} [depth] - 現在の再帰深度
- * @returns {Record<string, unknown>} サニタイズ済みの詳細情報
- */
-async function sanitizeLogDetails(
-    details: Record<string, unknown>,
-    visitedObjects?: WeakSet<object>,
-    depth = 0
-): Promise<Record<string, unknown>> {
-    // 入力チェック
-    if (details === null || details === undefined) {
-        return details;
-    }
-
-    if (typeof details !== 'object') {
-        throw new Error(`Expected object, got ${typeof details}`);
-    }
-
-    // 循環参照検出の初期化
-    if (typeof WeakSet !== 'undefined' && !visitedObjects) {
-        visitedObjects = new WeakSet<object>();
-    }
-
-    // 深度制限チェック
-    if (depth >= MAX_RECURSION_DEPTH) {
-        return { __sanitized: SANITIZE_RESULT.TOO_DEEP };
-    }
-
-    // 循環参照検出
-    if (visitedObjects && visitedObjects.has(details)) {
-        return { __sanitized: SANITIZE_RESULT.CIRCULAR_REF };
-    }
-
-    // メタオブジェクトは文字列化
-    if (details instanceof Date) {
-        return { __value: details.toISOString() };
-    }
-
-    if (details instanceof Error) {
-        return { message: details.message, stack: details.stack };
-    }
-
-    // WeakSetに現在のオブジェクトを追加
-    if (visitedObjects) {
-        visitedObjects.add(details);
-    }
-
-    const sanitized: Record<string, unknown> = {};
-
-    for (const [key, value] of Object.entries(details)) {
-        if (value === null || value === undefined) {
-            sanitized[key] = value;
-            continue;
-        }
-
-        // 文字列値の場合はPII検出を実行
-        if (typeof value === 'string') {
-            const result = await sanitizeRegex(value);
-            if (result.maskedItems.length > 0) {
-                sanitized[key] = result.text;
-                sanitized[`${key}_maskedTypes`] = result.maskedItems.map((m) => typeof m === 'string' ? m : m.type);
-            } else {
-                sanitized[key] = value;
-            }
-        } else if (typeof value === 'object') {
-            // 配列の明示的な処理
-            if (Array.isArray(value)) {
-                sanitized[key] = await sanitizeArray(value, visitedObjects, depth + 1);
-            } else {
-                // オブジェクトの場合は再帰的に処理
-                sanitized[key] = await sanitizeLogDetails(value as Record<string, unknown>, visitedObjects, depth + 1);
-            }
-        } else {
-            sanitized[key] = value; // primitive types
-        }
-    }
-
-    return sanitized;
-}
-
-/**
- * 配列を再帰的にサニタイズする（ヘルパー関数）
- */
-async function sanitizeArray(
-    arr: unknown[],
-    visitedObjects?: WeakSet<object>,
-    depth = 0
-): Promise<unknown[] | string> {
-    // 深度制限チェック
-    if (depth >= MAX_RECURSION_DEPTH) {
-        return SANITIZE_RESULT.TOO_DEEP;
-    }
-
-    // 循環参照検出
-    if (visitedObjects && visitedObjects.has(arr)) {
-        return SANITIZE_RESULT.CIRCULAR_REF;
-    }
-
-    // WeakSetに配列を追加
-    if (visitedObjects) {
-        visitedObjects.add(arr);
-    }
-
-    const sanitized: unknown[] = [];
-
-    for (const item of arr) {
-        if (item === null || item === undefined) {
-            sanitized.push(item);
-            continue;
-        }
-
-        if (typeof item === 'string') {
-            const result = await sanitizeRegex(item);
-            if (result.maskedItems.length > 0) {
-                sanitized.push(result.text);
-            } else {
-                sanitized.push(item);
-            }
-        } else if (typeof item === 'object') {
-            if (Array.isArray(item)) {
-                sanitized.push(await sanitizeArray(item, visitedObjects, depth + 1));
-            } else {
-                // メタオブジェクトのチェック
-                if (item instanceof Date) {
-                    sanitized.push(item.toISOString());
-                } else if (item instanceof Error) {
-                    sanitized.push({ message: item.message, stack: item.stack });
-                } else {
-                    sanitized.push(await sanitizeLogDetails(item as Record<string, unknown>, visitedObjects, depth + 1));
-                }
-            }
-        } else {
-            sanitized.push(item); // primitive types
-        }
-    }
-
-    return sanitized;
-}
-
-/**
- * Add a log entry
- * @param {LogTypeValues} type - LogType
- * @param {string} message - Log message
- * @param {object} [details] - Additional details (NO RAW PII)
- */
-export async function addLog<T extends object = Record<string, unknown>>(type: LogTypeValues, message: string, details: T = {} as T): Promise<void> {
-    try {
-        // 【セキュリティ強化】本番環境ではDEBUGログを破棄
-        // 【実装方針】: isDevelopment()で環境判定し、本番ならDEBUGを早期return
-        // 【テスト対応**: logger-production.test.ts - 本番環境のDEBUGログが出力されない
-        // 🟡 信頼性レベル: 黄信号（要件定義書のログ制約による）
-        if (!isDevelopment() && type === 'DEBUG') {
-            return; // DEBUGログは保存せず破棄
-        }
-
-        const sanitizedMessage = await sanitizeRegex(message);
-
-        const entry: LogEntry = {
-            id: typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function'
-                ? crypto.randomUUID()
-                : (() => { const a = new Uint32Array(2); crypto.getRandomValues(a); return a[0].toString(36) + a[1].toString(36); })(),
-            timestamp: Date.now(),
-            type,
-            message: sanitizedMessage.maskedItems.length > 0 ? sanitizedMessage.text : message,
-            details: await sanitizeLogDetails(details as Record<string, unknown>)
-        };
-
-        // バッファに追加（上限超過時は古いエントリを破棄）
-        if (pendingLogs.length >= MAX_PENDING_LOGS) {
-            // slice(1) creates new array but avoids in-place shifting
-            pendingLogs = pendingLogs.slice(1);
-        }
-        pendingLogs.push(entry);
-
-        // 【パフォーマンス改善】フラッシュ条件をチェック
-        if (pendingLogs.length >= BATCH_FLUSH_SIZE) {
-            await flushLogs();
-        } else {
-            scheduleFlush();
-        }
-    } catch (e) {
-        console.error('Logger: Failed to save log', e);
-    }
-}
-
-/**
- * Get all logs (including pending logs)
- * @returns {Promise<LogEntry[]>}
- */
-export async function getLogs(): Promise<LogEntry[]> {
-    const storage = await chrome.storage.local.get(LOG_STORAGE_KEY);
-    const storedLogs = (storage[LOG_STORAGE_KEY] as LogEntry[]) || [];
-    return [...storedLogs, ...pendingLogs];
-}
-
-/**
- * Clear all logs
- */
-export async function clearLogs(): Promise<void> {
-    pendingLogs = []; // 保留中のログもクリア
-    clearScheduledFlush();
-    await chrome.storage.local.remove(LOG_STORAGE_KEY);
-}
-
-/**
- * Filter out logs older than RETENTION_DAYS
- * @param {LogEntry[]} logs
- * @returns {LogEntry[]}
- */
-function pruneLogs(logs: LogEntry[]): LogEntry[] {
-    const cutoff = Date.now() - (RETENTION_DAYS * 24 * 60 * 60 * 1000);
-    return logs.filter(log => log.timestamp > cutoff);
-}
-
-// 【SRE/Logging改善 #8】構造化ロギング便利関数
-
-/**
- * Extract a short file name from an import.meta.url-style URL.
- * Strips query strings/fragments, removes the `.js` extension, and returns
- * the last path segment. Falls back to the original URL on parse errors.
- */
-export function extractSourceFromImportMetaUrl(url: string): string {
-    try {
-        const parsed = new URL(url);
-        const filename = parsed.pathname.split('/').pop() || 'unknown';
-        return filename.replace(/\.(js|ts)$/i, '') || 'unknown';
-    } catch {
-        const fallback = url.split('/').pop() || 'unknown';
-        return fallback.replace(/\.(js|ts)$/i, '') || 'unknown';
-    }
-}
-
-/**
- * Resolve the `source` field for a log entry.
- * If an explicit source is provided, it is returned unchanged.
- * Otherwise, the caller location is inferred from `Error.stack`.
+ * Implementation is split across logger/types.ts (error codes, log entry
+ * types), logger/core.ts (buffering, sanitization, addLog/getLogs/clearLogs),
+ * and logger/api.ts (logInfo/logWarn/logError/logDebug/logSanitize/logCritical).
+ * This file remains as a stable import path for the ~120 existing call sites.
  *
- * Note: after bundling, stack frames may point to chunk files rather than
- * original source files. The returned value is still useful for tracing but
- * may not match the TypeScript source name.
+ * @module logger
  */
-export function resolveLogSource(source?: string): string | undefined {
-    if (source !== undefined) {
-        return source;
-    }
+export {
+    ErrorCode,
+    LogType,
+} from './logger/types.js';
+export type {
+    ErrorCodeValues,
+    ErrorCodePattern,
+    LogTypeValues,
+    LogEntry,
+} from './logger/types.js';
 
-    const stack = new Error().stack || '';
-    const lines = stack.split('\n');
+export {
+    isDevelopment,
+    flushLogs,
+    getPendingLogCount,
+    clearPendingLogs,
+    addLog,
+    getLogs,
+    clearLogs,
+} from './logger/core.js';
 
-    for (const line of lines) {
-        // Skip frames that are part of this logger module.
-        if (line.includes('/logger.ts') || line.includes('\\logger.ts')) {
-            continue;
-        }
-
-        // Match common stack frame shapes:
-        //   at functionName (file:///path/to/file.ts:123:45)
-        //   at file:///path/to/file.ts:123:45
-        //   at functionName (/path/to/file.ts:123:45)
-        const match = line.match(/at\s+(?:[^\s(]+\s+)?\(?([^\s)]+?):\d+(?::\d+)?\)?$/);
-        if (match) {
-            return extractSourceFromImportMetaUrl(match[1]);
-        }
-    }
-
-    return 'unknown';
-}
-
-/**
- * 構造化されたログエントリを作成する（内部関数）
- * @param {LogTypeValues} type - ログタイプ
- * @param {string} message - メッセージ
- * @param {object} details - 詳細情報
- * @param {ErrorCodeValues} [errorCode] - エラーコード
- * @param {string} [source] - ログ出力元モジュール
- * @returns {LogEntry} ログエントリ
- */
-function createStructuredLog<T extends object = Record<string, unknown>>(
-    type: LogTypeValues,
-    message: string,
-    details: T = {} as T,
-    errorCode?: ErrorCodeValues,
-    source?: string
-): LogEntry {
-    return {
-        id: typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function' ? crypto.randomUUID() : Math.random().toString(36).substring(2),
-        timestamp: Date.now(),
-        type,
-        message,
-        errorCode,
-        source,
-        details: details as Record<string, unknown>
-    };
-}
-
-/**
- * 構造化されたINFOログを出力する
- * @param {string} message - メッセージ
- * @param {Record<string, unknown>} details - 詳細情報
- * @param {string} [source] - ログ出力元モジュール
- */
-export async function logInfo<T extends object = Record<string, unknown>>(
-    message: string,
-    details: T = {} as T,
-    source?: string
-): Promise<void> {
-    const entry = createStructuredLog(LogType.INFO, message, details, undefined, resolveLogSource(source));
-    await writeStructuredLog(entry);
-}
-
-/**
- * 構造化されたWARNログを出力する
- * @param {string} message - メッセージ
- * @param {object} details - 詳細情報
- * @param {ErrorCodeValues} [errorCode] - エラーコード
- * @param {string} [source] - ログ出力元モジュール
- */
-export async function logWarn<T extends object = Record<string, unknown>>(
-    message: string,
-    details: T = {} as T,
-    errorCode?: ErrorCodeValues,
-    source?: string
-): Promise<void> {
-    const entry = createStructuredLog(LogType.WARN, message, details, errorCode, resolveLogSource(source));
-    await writeStructuredLog(entry);
-}
-
-/**
- * 構造化されたERRORログを出力する
- * @param {string} message - メッセージ
- * @param {object} details - 詳細情報
- * @param {ErrorCodeValues} errorCode - エラーコード
- * @param {string} [source] - ログ出力元モジュール
- */
-export async function logError<T extends object = Record<string, unknown>>(
-    message: string,
-    details: T = {} as T,
-    errorCode: ErrorCodeValues = ErrorCode.UNKNOWN_ERROR,
-    source?: string
-): Promise<void> {
-    const entry = createStructuredLog(LogType.ERROR, message, details, errorCode, resolveLogSource(source));
-    await writeStructuredLog(entry);
-
-    // 開発環境ではconsole.errorにも出力
-    if (isDevelopment()) {
-        console.error(`[${errorCode}] ${message}`, details);
-    }
-}
-
-/**
- * 構造化されたDEBUGログを出力する
- * @param {string} message - メッセージ
- * @param {object} details - 詳細情報
- * @param {string} [source] - ログ出力元モジュール
- */
-export async function logDebug<T extends object = Record<string, unknown>>(
-    message: string,
-    details: T = {} as T,
-    source?: string
-): Promise<void> {
-    // 本番環境ではDEBUGログを出力しない
-    if (!isDevelopment()) {
-        return;
-    }
-    const entry = createStructuredLog(LogType.DEBUG, message, details, undefined, resolveLogSource(source));
-    await writeStructuredLog(entry);
-
-    // 開発環境ではconsole.debugにも出力
-    if (isDevelopment()) {
-        console.debug(`[DEBUG] ${message}`, details);
-    }
-}
-
-/**
- * 構造化されたSANITIZEログを出力する
- * @param {string} message - メッセージ
- * @param {object} details - 詳細情報
- * @param {ErrorCodeValues} [errorCode] - エラーコード
- * @param {string} [source] - ログ出力元モジュール
- */
-export async function logSanitize<T extends object = Record<string, unknown>>(
-    message: string,
-    details: T = {} as T,
-    errorCode?: ErrorCodeValues,
-    source?: string
-): Promise<void> {
-    const entry = createStructuredLog(LogType.SANITIZE, message, details, errorCode, resolveLogSource(source));
-    await writeStructuredLog(entry);
-}
-
-/**
- * 構造化ログを書き込む（内部関数）
- * @param {LogEntry} entry - ログエントリ
- */
-async function writeStructuredLog(entry: LogEntry): Promise<void> {
-    try {
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const { id, timestamp, type, message, errorCode, source, details } = entry;
-
-        // 既存のaddLog関数を使用（entryを分割して渡す）
-        await addLog(
-            type,
-            message,
-            {
-                ...details,
-                _errorCode: errorCode,
-                _source: source
-            }
-        );
-    } catch (e) {
-        console.error('Logger: Failed to write structured log', e);
-    }
-}
-
-const CRITICAL_NOTIFICATION_COOLDOWN_MS = 5 * 60 * 1000;
-let lastCriticalNotificationTime = 0;
-
-/**
- * CRITICAL — 構造化ERRORログ + chrome.notifications アラート
- * 暗号化失敗、データ損失リスクなど重大な障害で使用する。
- * 通知にはクールダウン（5分）があり、連続通知スパムを防ぐ。
- */
-export async function logCritical<T extends object = Record<string, unknown>>(
-    message: string,
-    details: T = {} as T,
-    errorCode: ErrorCodeValues = ErrorCode.UNKNOWN_ERROR,
-    source?: string
-): Promise<void> {
-    const entry = createStructuredLog(LogType.ERROR, message, details, errorCode, resolveLogSource(source));
-    await writeStructuredLog(entry);
-    // Critical logs are flushed immediately so they are not lost on SW termination.
-    await flushLogs(true);
-
-    console.error(`[CRITICAL:${errorCode}] ${message} ${JSON.stringify(details, (key, value) => {
-        if (typeof value === 'string' && value.length > 128) {
-            return value.slice(0, 128) + '...[truncated]';
-        }
-        if (typeof value === 'string' && /^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$/.test(value) && value.length > 40) {
-            return value.slice(0, 8) + '...[redacted]';
-        }
-        return value;
-    })}`);
-
-    const now = Date.now();
-    if (now - lastCriticalNotificationTime < CRITICAL_NOTIFICATION_COOLDOWN_MS) {
-        return;
-    }
-    lastCriticalNotificationTime = now;
-
-    try {
-        if (typeof chrome !== 'undefined' && chrome.notifications && typeof chrome.notifications.create === 'function') {
-            const title = chrome.i18n.getMessage('criticalAlertTitle') || 'Yasumaro — Critical Error';
-            const notificationMessage = chrome.i18n.getMessage('criticalAlertBody', [message]) || message;
-            const iconUrl = (typeof chrome.runtime !== 'undefined' && typeof chrome.runtime.getURL === 'function')
-                ? chrome.runtime.getURL('icons/icon48.png')
-                : 'icons/icon48.png';
-
-            chrome.notifications.create({
-                type: 'basic',
-                iconUrl,
-                title,
-                message: notificationMessage,
-                priority: 2,
-                requireInteraction: true,
-            });
-        }
-    } catch (e) {
-        console.error('Logger: Failed to create critical notification', e);
-    }
-}
+export {
+    extractSourceFromImportMetaUrl,
+    resolveLogSource,
+    logInfo,
+    logWarn,
+    logError,
+    logDebug,
+    logSanitize,
+    logCritical,
+} from './logger/api.js';
